Show just-created Circular after new submission
Fixes #1699.
